requirements.txt
Flask==2.3.3
Pillow==10.0.1
pillow-avif-plugin==1.4.3